update long to long long

+++ Amplitude/AMPDatabaseHelper.h
@@ -17,17 +17,17 @@
 
 - (BOOL)addEvent:(NSString*) event;
 - (BOOL)addIdentify:(NSString*) identify;
-- (NSMutableArray*)getEvents:(long) upToId limit:(long) limit;
-- (NSMutableArray*)getIdentifys:(long) upToId limit:(long) limit;
+- (NSMutableArray*)getEvents:(long long) upToId limit:(long long) limit;
+- (NSMutableArray*)getIdentifys:(long long) upToId limit:(long long) limit;
 - (int)getEventCount;
 - (int)getIdentifyCount;
 - (int)getTotalEventCount;
-- (BOOL)removeEvents:(long) maxId;
-- (BOOL)removeIdentifys:(long) maxIdentifyId;
-- (BOOL)removeEvent:(long) eventId;
-- (BOOL)removeIdentify:(long) identifyId;
-- (long long)getNthEventId:(long) n;
-- (long long)getNthIdentifyId:(long) n;
+- (BOOL)removeEvents:(long long) maxId;
+- (BOOL)removeIdentifys:(long long) maxIdentifyId;
+- (BOOL)removeEvent:(long long) eventId;
+- (BOOL)removeIdentify:(long long) identifyId;
+- (long long)getNthEventId:(long long) n;
+- (long long)getNthIdentifyId:(long long) n;
 
 - (BOOL)insertOrReplaceKeyValue:(NSString*) key value:(NSString*) value;
 - (BOOL)insertOrReplaceKeyLongValue:(NSString*) key value:(NSNumber*) value;
